Use registry as part of the repository
Before the big 1.0.0 pull, we had an additional value for the registry. I feel
this isn't really needed and also causes a bit more difficulty when trying to
run tests (images may not be in quay.io, for example when checking a newer
version).

SO go back to regular repository name with the registry prepended.