kong: refactor env block generation
Significantly refactor the generation of Kong container env blocks. Add
both user-provided variables and variables generated from other sections
of values.yaml to dictionaries, merge the dictionaries to ensure only a
single instance of a variable is present, and render the resulting
dictionary into an EnvVar block. In addition:

* Condense kong.env and kong.final_env back into a single kong.env template.
* Remove special-case env generation from kong.wait-for-db and use
  kong.env to ensure consistent behavior accross all Kong containers.
* Add a new kong.no_daemon_env template to disable NGINX daemonization for
  containers other than wait-for-db.
* DRY license, session configuration, and auth configuration templates into
  a generic "secretkeyref" template.
* Add a new case to the renderer for handling templated secretKeyRef
  blocks. Templates can only generate strings, not string maps, and the
  quote function condenses multi-line strings onto a single line, so
  neither of the existing render cases work.

The additional renderer relies on the presence of "valueFrom" as a magic
string to indicate that it should handle that variable. It cannot handle
other multi-line blocks as such. At present (API version 1.17),
valueFrom is the only such block suppported by the EnvVar specification.

Signed-off-by: Travis Raines <traines@konghq.com>

From #4