[kong] Don't recreate service account (#31)
Remove the temporary ServiceAccount created via hook during pre-upgrade
migrations. This was previously in place to automatically handle users
upgrading from a configuration that did not use a ServiceAccount to one that
did. However, this temporary account overwrites credentials for the real
account if left in the template.

Add instructions for temporarily modifying the template to re-add this hook to
the FAQ, along with examples of the error that will appear.