[kong] use internal address for admin API when appropriate (#125)
Use 127.0.0.1 as the admin listen address if the admin service is
disabled. This is primarily for controller-managed deployments, where
the admin API should not be exposed outside the pod (removing it from
the deployment ports alone does not prevent this).

This changes adds generic support for specifying the listen address, but
does not use local listens for other services: generally, the admin API
is the only service we expect to encounter a disabled service with
enabled listens.

If SVC.enabled=false, do not render ports for that service in the
deployment template.