[kong] use bash wait-for-postgres implementation (#179)
Replace wait-for-postgres netcat invocation with an equivalent bash
script. This removes the dependency for a specific waitImage, which is
necessary for environments where busybox is not available (namely in the
context of Red Hat certified operators, which must use Red Hat images
for all containers).

Although this removes the busybox (and netcat) dependency, it does
introduce a bash dependency. Because bash is not available on busybox,
centos is now the standard waitImage.

The bash script uses a static ConfigMap to work around some issues with
newline handling in Job declarations.